modified:   backend/controllers/workoutController.js 	modified:   backend/middleware/requireAuth.js 	modified:   backend/models/workoutModel.js 	modified:   frontend/src/App.js 	modified:   frontend/src/components/WorkoutDetails.js 	modified:   frontend/src/components/WorkoutForm.js 	modified:   frontend/src/contexts/AuthContext.js 	modified:   frontend/src/hooks/useLogout.js 	modified:   frontend/src/index.css 	modified:   frontend/src/pages/Home.js

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,22 @@
 import { useContext } from "react";
 import { WorkoutsContext } from "../contexts/WorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({workout}) => {
     const {dispatch} = useContext(WorkoutsContext);
 
+    const {user} = useAuthContext();
+
     const handleClick = async () =>{
+        if(!user){
+            return;
+        }
+
         const response = await fetch('/api/workouts/'+workout._id ,{
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization' : `Bearer ${user.jwt}`
+            }
         });
         const json = await response.json();
